fix for question 10 func, and its tests

+++ src/test/java/dSoft/Bloco2/Ex2DigitsTest.java
@@ -9,14 +9,14 @@ import static org.junit.Assert.*;
 public class Ex2DigitsTest {
 	@Test
 	public void validLength() {
-		int providedNum = 123;
+		byte providedNum = 123;
 		boolean result = DigitMethods.verifyLength(providedNum);
 		assertTrue(result);
 	}
 
 	@Test
 	public void validNumberDigits() {
-		int providedNum = 123;
+		byte providedNum = 123;
 		int[] expected = {1, 2, 3};
 		int[] result = DigitMethods.getDigits(providedNum);
 		assertArrayEquals(expected, result);
@@ -24,14 +24,14 @@ public class Ex2DigitsTest {
 
 	@Test
 	public void invalidLength() {
-		int providedNum = 5;
+		byte providedNum = 5;
 		boolean result = DigitMethods.verifyLength(providedNum);
 		assertFalse(result);
 	}
 
 	@Test
 	public void invalidInput() {
-		int providedNum = 5;
+		byte providedNum = 5;
 		int[] expected = {-1};
 		int[] result = DigitMethods.getDigits(providedNum);
 		assertArrayEquals(expected, result);
